+++ .pre-commit-config.yaml
@@ -18,7 +18,7 @@ repos:
     rev: 6.0.0
     hooks:
       - id: flake8
-        args: [ "--max-line-length=120", "--ignore=F401,W503,E203,E731,F841,E402" ]
+        args: [ "--max-line-length=120", "--ignore=F401,W503,E203,E731,F841,E402,F811" ]
 
   - repo: local
     hooks:
