+++ chalicelib/domain/brand/model/brand_response.py
@@ -2,11 +2,8 @@ import re
 
 from marshmallow import (
     EXCLUDE,
-    INCLUDE,
     Schema,
-    ValidationError,
     fields,
-    validates_schema,
 )
 
 
@@ -36,33 +33,16 @@ class _BrandDetailEventResponseSchema(Schema):
         unknown = EXCLUDE
 
 
-class _BrandDetailProductItemResponseSchema(Schema):
+class _BrandDetailProductResponseSchema(Schema):
     id = fields.Int(required=True)
     name = fields.Str(required=True)
     image = fields.URL(required=True, allow_none=True)
     image_alt = fields.Str(required=True)
     price = fields.Float(required=True)
-
-
-class _BrandDetailProductResponseSchema(Schema):
-    events = fields.List(fields.Str(), required=True)
-
-    @validates_schema
-    def validate_items(self, data, **kwargs):
-        item_schema = _BrandDetailProductItemResponseSchema()
-        events = data["events"]
-        for event in events:
-            if event not in data or len(data[event]) == 0:
-                raise ValidationError(f"Event {event} not found in data", event)
-            categories = data[event].keys()
-            for category in categories:
-                errors = item_schema.validate(data[event][category], many=True)
-                if errors:
-                    raise ValidationError("\n".join(errors), f"{event}.{category}")
+    good_count: int = fields.Integer(required=True)
 
     class Meta:
-        order = True
-        unknown = INCLUDE
+        unknown = EXCLUDE
 
 
 class BrandDetailResponseSchema(Schema):
@@ -72,7 +52,9 @@ class BrandDetailResponseSchema(Schema):
     meta = fields.Nested(_BrandDetailMetaResponseSchema, required=True)
     description = fields.Str(required=True)
     events = fields.Nested(_BrandDetailEventResponseSchema, many=True, required=True)
-    products = fields.Nested(_BrandDetailProductResponseSchema, required=True)
+    products = fields.Nested(
+        _BrandDetailProductResponseSchema, required=True, many=True
+    )
 
     class Meta:
         ordered = True
